Feature/WD-51-web-token (#44)
* rozgrzebane xd

* guests doneeeee

* decorators cleanup

* budget update first version

* providers and dump

* Dump poprawki

* porządek z gośćmi

* porządek z resztą

* category limits fixes

* final version

* After master merge

* tokens adjustments

* rm mobile and e2e

* conf

* dostosowanie frontu providerów do endpointów

* Exp date

* Additional info about user in /profile

* Categories in swagger

* on errors and on unauthorized

* delete providers and category fix

* language

* database fix

* translations, login screens swap

* Del statusBar

* language adjustments

---------

+++ apps/backend/src/app/modules/providers/services/providers.service.ts
@@ -14,17 +14,17 @@ import { PrismaService } from '../../../prisma-client.service';
 export class ProvidersService {
   constructor(private prisma: PrismaService) {}
 
-  async createNewCategory(newCategoryDto: CreateProviderCategoryDto): Promise<ProviderCategoryDto> {
-    return this.prisma.providerCategory.create({ data: newCategoryDto });
+  async createNewCategory(userId: number, newCategoryDto: CreateProviderCategoryDto): Promise<ProviderCategoryDto> {
+    return this.prisma.providerCategory.create({ data: { ...newCategoryDto, userId } });
   }
 
-  async removeCategory(id: number): Promise<ProviderCategoryDto> {
+  async removeCategory(userId: number, categoryId: number): Promise<ProviderCategoryDto> {
     try {
       await this.prisma.provider.deleteMany({
-        where: { categoryId: id },
+        where: { categoryId, userId },
       });
 
-      return this.prisma.providerCategory.delete({ where: { id: id } });
+      return this.prisma.providerCategory.delete({ where: { id: categoryId, userId } });
     } catch (e) {
       if (e.code === 'P2025') {
         throw new Error('Provider category not found');
@@ -33,9 +33,9 @@ export class ProvidersService {
     }
   }
 
-  async getCategoryById(id: number): Promise<ProviderCategoryDto> {
+  async getCategoryById(userId: number, id: number): Promise<ProviderCategoryDto> {
     try {
-      return this.prisma.providerCategory.findUnique({ where: { id: id } });
+      return this.prisma.providerCategory.findUnique({ where: { id, userId } });
     } catch (e) {
       if (e.code === 'P2025') {
         throw new Error('Provider category not found');
@@ -44,7 +44,7 @@ export class ProvidersService {
     }
   }
 
-  private async getAllCategoriesWithProviders(): Promise<CategoryWithProviders[]> {
+  private async getAllCategoriesWithProviders(userId: number): Promise<CategoryWithProviders[]> {
     return this.prisma.providerCategory.findMany({
       include: {
         providers: {
@@ -54,11 +54,12 @@ export class ProvidersService {
           },
         },
       },
+      where: { userId },
     });
   }
 
-  async getCategoriesSummary(): Promise<CategoryToSummaryDto[]> {
-    const categories = await this.getAllCategoriesWithProviders();
+  async getCategoriesSummary(userId: number): Promise<CategoryToSummaryDto[]> {
+    const categories = await this.getAllCategoriesWithProviders(userId);
     return categories.map((category) => {
       const totalProviders: number = category.providers.length;
       const reservedProviders: number = category.providers.filter((provider) => provider.isReserved).length;
@@ -73,14 +74,14 @@ export class ProvidersService {
     });
   }
 
-  async createNewProvider(newProviderDto: CreateProviderDto): Promise<ProviderDto> {
-    return this.prisma.provider.create({ data: newProviderDto });
+  async createNewProvider(userId: number, newProviderDto: CreateProviderDto): Promise<ProviderDto> {
+    return this.prisma.provider.create({ data: { ...newProviderDto, userId } });
   }
 
-  async updateProvider(id: number, updateProviderDto: UpdateProviderDto): Promise<ProviderDto> {
+  async updateProvider(userId: number, id: number, updateProviderDto: UpdateProviderDto): Promise<ProviderDto> {
     try {
       return await this.prisma.provider.update({
-        where: { id },
+        where: { id, userId },
         data: updateProviderDto,
       });
     } catch (e) {
@@ -91,20 +92,20 @@ export class ProvidersService {
     }
   }
 
-  async getAllProvidersInCategory(categoryId: number): Promise<ProviderDto[]> {
+  async getAllProvidersInCategory(userId: number, categoryId: number): Promise<ProviderDto[]> {
     try {
       return await this.prisma.provider.findMany({
-        where: { categoryId },
+        where: { categoryId, userId },
       });
     } catch (e) {
       throw new Error(e);
     }
   }
 
-  async getProviderById(id: number): Promise<ProviderDto> {
+  async getProviderById(userId: number, id: number): Promise<ProviderDto> {
     try {
       return await this.prisma.provider.findFirst({
-        where: { id },
+        where: { id, userId },
       });
     } catch (e) {
       if (e.code === 'P2025') {
@@ -114,9 +115,9 @@ export class ProvidersService {
     }
   }
 
-  async removeProvider(id: number): Promise<ProviderDto> {
+  async removeProvider(userId: number, id: number): Promise<ProviderDto> {
     try {
-      return await this.prisma.provider.delete({ where: { id } });
+      return await this.prisma.provider.delete({ where: { id, userId } });
     } catch (e) {
       if (e.code === 'P2025') {
         throw new Error('Provider not found');
@@ -125,9 +126,12 @@ export class ProvidersService {
     }
   }
 
-  async getProvidersGroupedByStarsForCategory(categoryId: number): Promise<{ title: string; data: ProviderDto[] }[]> {
+  async getProvidersGroupedByStarsForCategory(
+    userId: number,
+    categoryId: number
+  ): Promise<{ title: string; data: ProviderDto[] }[]> {
     const providersInCategory = await this.prisma.provider.findMany({
-      where: { categoryId },
+      where: { categoryId, userId },
     });
 
     const groupedProviders = providersInCategory.reduce(
